pyproject.toml
[tool.poetry]
name = "dating-app"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "dating_app"}]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.94.0"
uvicorn = {extras = ["standard"], version = "^0.21.0"}
oauth2 = "^1.9.0.post1"
python-multipart = "^0.0.6"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
